resolve nits
