+++ packages/make-styles/src/runtime/compileCSS.ts
@@ -1,5 +1,4 @@
-import { compile, middleware, serialize, stringify } from 'stylis';
-import { compileCSSRule } from './compileStaticCSS';
+import { compile, middleware, prefixer, rulesheet, serialize, stringify } from 'stylis';
 import { hyphenateProperty } from './utils/hyphenateProperty';
 
 interface CompileCSSOptions {
@@ -17,6 +16,29 @@ interface CompileCSSOptions {
 
 function repeatSelector(selector: string, times: number) {
   return new Array(times + 2).join(selector);
+}
+
+export function compileCSSRule(cssRule: string): string {
+  return serialize(compile(cssRule), middleware([prefixer, stringify]));
+}
+
+export function compileCSSRules(cssRules: string): string[] {
+  const rules: string[] = [];
+
+  serialize(
+    compile(cssRules),
+    middleware([
+      prefixer,
+      stringify,
+
+      // 💡 we are using `.insertRule()` API for DOM operations, which does not support
+      // insertion of multiple CSS rules in a single call. `rulesheet` plugin extracts
+      // individual rules to be used with this API
+      rulesheet(rule => rules.push(rule)),
+    ]),
+  );
+
+  return rules;
 }
 
 export function compileCSS(options: CompileCSSOptions): string {
